+++ src/main/java/dev/lambdaurora/mcdev/LambdaMcDevGradlePlugin.java
@@ -8,17 +8,26 @@
 
 package dev.lambdaurora.mcdev;
 
+import dev.lambdaurora.mcdev.api.LambdaMcDevGradleExtension;
+import dev.lambdaurora.mcdev.ext.LambdaMcDevGradleExtensionImpl;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
+import org.gradle.api.plugins.JavaPluginExtension;
+import org.gradle.api.reflect.TypeOf;
 import org.jetbrains.annotations.NotNull;
 
 public class LambdaMcDevGradlePlugin implements Plugin<Project> {
 	@Override
 	public void apply(@NotNull Project project) {
-		// Apply loom.
-		project.getPlugins().apply("fabric-loom");
-
-		var loomExt = project.getExtensions().getByName("loom");
-		project.getExtensions().create("lambdamcdev", LambdaMcDevGradleExtension.class, loomExt);
+		var javaExt = project.getExtensions().getByType(JavaPluginExtension.class);
+		var loomExt = project.getExtensions().findByName("loom");
+		project.getExtensions().create(
+				TypeOf.typeOf(LambdaMcDevGradleExtension.class),
+				"lambdamcdev",
+				LambdaMcDevGradleExtensionImpl.class,
+				project,
+				javaExt,
+				loomExt
+		);
 	}
 }