+++ CHANGELOG.md
@@ -5,3 +5,9 @@
 - Initial release.
 - Added support YALMM mappings.
 - Added utilities related to versioning.
+
+## 1.1.0
+
+- Added tasks to generate `fabric.mod.json` and `neoforge.mods.toml` files.
+- Added experimental access widener to access transformer converter.
+- Tweaked how loom is required to be compatible with loom forks.